add issue reproducer
The issue is caused by the test class being loaded by a different
classloader than its superclass. Specifically, the test class is
loaded by `Quarkus Base Runtime ClassLoader`, but its parent class
is loaded by the system classloader. This class uses RestAssured,
so that is loaded by the system classloader as well. And now we have
RestAssured loaded by 2 classloaders -- the Quarkus one, which has
the correct URL and port set, and the system classloader one, which
doesn't have the correct settings, but is the one that is actually used.